+++ elmyra.ip.access.epo/CHANGES.rst
@@ -8,6 +8,7 @@ development
 - ui: compute cql query from comfort form fields already when switching tabs
 - ui: query builder action button reorganization
 - ui: basket action button reorganization
+- ui: project action button reorganization
 
 0.35.0
 ======